+++ worlds/stardew_valley/data/locations.csv
@@ -2252,7 +2252,7 @@ id,region,name,tags,mod_name
 3848,Shipping,Shipsanity: Way Of The Wind pt. 1,"SHIPSANITY",
 3849,Shipping,Shipsanity: Mapping Cave Systems,"SHIPSANITY",
 3850,Shipping,Shipsanity: Price Catalogue,"SHIPSANITY",
-3851,Shipping,Shipsanity: Queen Of Sauce Cookbook,"SHIPSANITY",
+3851,Shipping,Shipsanity: Queen Of Sauce Cookbook,"SHIPSANITY,GINGER_ISLAND",
 3852,Shipping,Shipsanity: The Diamond Hunter,"SHIPSANITY,GINGER_ISLAND",
 3853,Shipping,Shipsanity: Book of Mysteries,"SHIPSANITY",
 3854,Shipping,Shipsanity: Animal Catalogue,"SHIPSANITY",
@@ -2292,7 +2292,7 @@ id,region,name,tags,mod_name
 4032,Farm,Read Book Of Stars,"BOOKSANITY,BOOKSANITY_SKILL",
 4033,Farm,Read Combat Quarterly,"BOOKSANITY,BOOKSANITY_SKILL",
 4034,Farm,Read Mining Monthly,"BOOKSANITY,BOOKSANITY_SKILL",
-4035,Farm,Read Queen Of Sauce Cookbook,"BOOKSANITY,BOOKSANITY_SKILL",
+4035,Farm,Read Queen Of Sauce Cookbook,"BOOKSANITY,BOOKSANITY_SKILL,GINGER_ISLAND",
 4036,Farm,Read Stardew Valley Almanac,"BOOKSANITY,BOOKSANITY_SKILL",
 4037,Farm,Read Woodcutter's Weekly,"BOOKSANITY,BOOKSANITY_SKILL",
 4051,Museum,Read Tips on Farming,"BOOKSANITY,BOOKSANITY_LOST",
