+++ worlds/stardew_valley/data/requirement.py
@@ -29,3 +29,8 @@ class SeasonRequirement(Requirement):
 @dataclass(frozen=True)
 class YearRequirement(Requirement):
     year: int
+
+
+@dataclass(frozen=True)
+class WalnutRequirement(Requirement):
+    amount: int
