+++ worlds/stardew_valley/logic/requirement_logic.py
@@ -9,8 +9,9 @@ from .season_logic import SeasonLogicMixin
 from .skill_logic import SkillLogicMixin
 from .time_logic import TimeLogicMixin
 from .tool_logic import ToolLogicMixin
+from .walnut_logic import WalnutLogicMixin
 from ..data.game_item import Requirement
-from ..data.requirement import ToolRequirement, BookRequirement, SkillRequirement, SeasonRequirement, YearRequirement
+from ..data.requirement import ToolRequirement, BookRequirement, SkillRequirement, SeasonRequirement, YearRequirement, WalnutRequirement
 
 
 class RequirementLogicMixin(BaseLogicMixin):
@@ -20,7 +21,7 @@ class RequirementLogicMixin(BaseLogicMixin):
 
 
 class RequirementLogic(BaseLogic[Union[RequirementLogicMixin, HasLogicMixin, ReceivedLogicMixin, ToolLogicMixin, SkillLogicMixin, BookLogicMixin,
-SeasonLogicMixin, TimeLogicMixin]]):
+SeasonLogicMixin, TimeLogicMixin, WalnutLogicMixin]]):
 
     def meet_all_requirements(self, requirements: Iterable[Requirement]):
         if not requirements:
@@ -50,3 +51,7 @@ SeasonLogicMixin, TimeLogicMixin]]):
     @meet_requirement.register
     def _(self, requirement: YearRequirement):
         return self.logic.time.has_year(requirement.year)
+
+    @meet_requirement.register
+    def _(self, requirement: WalnutRequirement):
+        return self.logic.walnut.has_walnut(requirement.amount)
